% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{best.ml.method.parameters}
\alias{best.ml.method.parameters}
\title{returns the two parameters that are used to find the best ML method}
\usage{
best.ml.method.parameters(
  BLP.obj,
  GATES.obj,
  proxy.cate.main.sample,
  group.membership.main.sample
)
}
\arguments{
\item{BLP.obj}{an object as returned by CLAN()}

\item{GATES.obj}{an object as returned by get.BLP.parameters()}

\item{proxy.cate.main.sample}{Proxy CATE estimators for the main sample}

\item{group.membership.main.sample}{a logical matrix with \emph{M} rows that indicate
the group memberships (such a matrix is returned by the function quantile.group())}
}
\value{
lambda and lambda.bar parameters
}
\description{
returns the two parameters that are used to find the best ML method
}
