% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{CATE.proxy.estimator}
\alias{CATE.proxy.estimator}
\title{Estimates the CATE proxy estimator \eqn{E[Y | D=1, Z] - E[Y | D=0, Z]} on the auxiliary sample}
\usage{
CATE.proxy.estimator(
  Z,
  D,
  Y,
  auxiliary.sample,
  learner = "random.forest",
  proxy.baseline.estimates = NULL,
  min_variation = 1e-05
)
}
\arguments{
\item{Z}{an ( \emph{n} x \emph{d}) matrix or data frame of covariates}

\item{D}{a binary vector of treatment status of length \emph{n}}

\item{Y}{a vector of responses of length \emph{n}}

\item{auxiliary.sample}{a numerical vector of indices of observations in the auxiliary sample. Length is shorter than \emph{n}}

\item{learner}{the regression machine learner to be used. Either 'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_g = mlr3::lrn("regr.ranger", num.trees = 500) for a regression forest, which is also the default.}

\item{proxy.baseline.estimates}{A vector of length \emph{n} of proxy estimates of the baseline estimator \eqn{E[Y | D=0, Z]}. If NULL, these will be estimated separately.}

\item{min_variation}{minimum variation of the predictions before random noise with distribution N(0, var(Y)/20) is added. Default is 1e-05.}
}
\value{
Estimates of the CATE, both for the auxiliary sample and all observations, and an 'mlr3' object of each employed model
}
\description{
Estimates the CATE proxy estimator \eqn{E[Y | D=1, Z] - E[Y | D=0, Z]} on the auxiliary sample
}
