% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clan.R
\name{CLAN}
\alias{CLAN}
\title{Estimates the CLAN parameters.}
\usage{
CLAN(
  Z_CLAN,
  membership,
  equal_variances = FALSE,
  diff = setup_diff(),
  significance_level = 0.05
)
}
\arguments{
\item{Z_CLAN}{A matrix of variables that shall be considered for the CLAN. Each column represents a variable for which CLAN shall be performed.}

\item{membership}{A logical matrix that indicates the group membership of each observation in \code{Z_CLAN}. Needs to be an instance of \code{\link{quantile_group}}.}

\item{equal_variances}{Logical. If \code{TRUE}, the the two within-group variances of the differences between the CLAN generic targets are assumed to be equal. Default is \code{FALSE}.}

\item{diff}{Specifies the generic targets of CLAN See the documentation of \code{\link{setup_diff}} for details.}

\item{significance_level}{Significance level. Default is 0.05.}
}
\value{
An object of the class \code{CLAN}.
}
\description{
Estimates the parameters of the Classification Analysis (CLAN) on all variables in the supplied design matrix \code{Z_CLAN}.
}
