% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{GATES}
\alias{GATES}
\title{Performs GATES regression.}
\usage{
GATES(
  Y,
  D,
  propensity_scores,
  proxy_BCA,
  proxy_CATE,
  membership,
  HT = FALSE,
  X1_control = setup_X1(),
  vcov_control = setup_vcov(),
  diff = setup_diff(),
  significance_level = 0.05
)
}
\arguments{
\item{Y}{A vector of responses.}

\item{D}{A binary vector of treatment status.}

\item{propensity_scores}{A vector of propensity scores.}

\item{proxy_BCA}{A vector of proxy baseline estimates.}

\item{proxy_CATE}{A vector of proxy CATE estimates.}

\item{membership}{A logical matrix that indicates the group membership of each observation in \code{Z_CLAN}. Needs to be an instance of \code{\link{quantile_group}}.}

\item{HT}{Logical. If \code{TRUE}, a HT transformation is applied (GATES2 in the paper). Default is \code{FALSE}.}

\item{X1_control}{Specifies the design matrix \eqn{X_1} in the regression. See the documentation of \code{\link{setup_X1}} for details.}

\item{vcov_control}{Specifies the covariance matrix estimator. See the documentation of \code{\link{setup_vcov}} for details.}

\item{diff}{Specifies the generic targets of GATES. See the documentation of \code{\link{setup_diff}} for details.}

\item{significance_level}{Significance level. Default is 0.05.}
}
\value{
An object of the class \code{GATES}.
}
\description{
Performs the linear regression for the GATES procedure.
}
