% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.generic.ml.for.given.learner}
\alias{get.generic.ml.for.given.learner}
\title{Performs generic ML for a given learning technique (with only one split of the data)}
\usage{
get.generic.ml.for.given.learner(
  Z,
  D,
  Y,
  propensity.scores,
  learner = "mlr3::lrn(\\"cv_glmnet\\", s = \\"lambda.min\\")",
  M.set,
  A.set,
  Z_CLAN = NULL,
  X1_BLP = setup_X1(),
  X1_GATES = setup_X1(),
  HT = FALSE,
  vcov_BLP = setup_vcov(),
  vcov_GATES = setup_vcov(),
  equal_variances_CLAN = FALSE,
  quantile_cutoffs = c(0.25, 0.5, 0.75),
  diff_GATES = setup_diff(),
  diff_CLAN = setup_diff(),
  significance_level = 0.05,
  min_variation = 1e-05
)
}
\arguments{
\item{Z}{a matrix or data frame of covariates}

\item{D}{a binary vector of treatment status of length}

\item{Y}{a vector of responses of length}

\item{propensity.scores}{a vector of propensity scores}

\item{learner}{The machine learner that shall be used}

\item{M.set}{main set}

\item{A.set}{auxiliary set}

\item{Z_CLAN}{A matrix of variables that shall be considered for the CLAN. If \code{NULL} (default), then \code{Z_CLAN = Z}, i.e. CLAN is performed for all variables in \code{Z}.}

\item{X1_BLP}{Specifies the design matrix \eqn{X_1} in the BLP regression. See the documentation of \code{\link{setup_X1}} for details.}

\item{X1_GATES}{Same as \code{X1_BLP}, just for the the GATES regression.}

\item{HT}{logical. If TRUE, a HT transformation is applied in BLP and GATES. Default is FALSE.}

\item{vcov_BLP}{Specifies the covariance matrix estimator in the BLP regression. See the documentation of \code{\link{setup_vcov}} for details.}

\item{vcov_GATES}{Same as \code{vcov_BLP}, just for the GATES regression.}

\item{equal_variances_CLAN}{logical. If TRUE, the the two within-group variances of the most and least affected group in CLAN are assumed to be equal. Default is FALSE.}

\item{quantile_cutoffs}{Cutoff points of quantiles that shall be used for GATES grouping}

\item{diff_GATES}{Specifies the generic targets of GATES. See the documentation of \code{\link{setup_diff}}.}

\item{diff_CLAN}{Same as \code{diff_GATES}, just for the CLAN generic targets.}

\item{significance_level}{Significance level. Default is 0.05}

\item{min_variation}{minimum variation of the predictions before random noise with distribution N(0, var(Y)/20) is added. Default is 1e-05.

TODO: instructions on how mlr3 input is supposed to work (needs to be a string!)
TODO: comments on CLAN: If there are categorical variables, apply one-hot-encoding to Z_CLAN. The interpretation then becomes: Is there a factor that is overproportionally present in the least or most affected group?}
}
\description{
Performs generic ML for a given learning technique (with only one split of the data)
}
