% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_BLP}
\alias{get_BLP}
\title{Accessor function for the BLP generic target estimates after VEIN}
\usage{
get_BLP(x, learner = "best", plot = TRUE)
}
\arguments{
\item{x}{An object of the class \code{\link{GenericML}}.}

\item{learner}{A character string of the learner whose BLP generic target estimates shall be accessed. Default is \code{"best"} for the best learner for BLP.}

\item{plot}{Logical. If \code{TRUE} (default), a plot is printed.}
}
\value{
A numeric matrix of BLP generic target estimates. Furthermore, prints a plot if \code{plot = TRUE}.
}
\description{
Accessor function for the BLP generic target estimates after VEIN
}
