% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_CLAN}
\alias{get_CLAN}
\title{Accessor function for the CLAN generic target estimates after VEIN}
\usage{
get_CLAN(x, variable, learner = "best", plot = TRUE)
}
\arguments{
\item{x}{An object of the class \code{\link{GenericML}}.}

\item{variable}{The (character) name of a variabe on which CLAN was performed.}

\item{learner}{A character string of the learner whose CLAN generic target estimates shall be accessed. Default is \code{"best"} for the best learner for CLAN}

\item{plot}{Logical. If \code{TRUE} (default), a plot is printed.}
}
\value{
A numeric matrix of CLAN generic target estimates. Furthermore, prints a plot if \code{plot = TRUE}.
}
\description{
Accessor function for the CLAN generic target estimates after VEIN
}
