% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_GATES}
\alias{get_GATES}
\title{Accessor function for the GATES generic target estimates after VEIN}
\usage{
get_GATES(x, learner = "best", plot = TRUE)
}
\arguments{
\item{x}{An object of the class \code{\link{GenericML}}.}

\item{learner}{A character string of the learner whose GATES generic target estimates shall be accessed. Default is \code{"best"} for the best learner for GATES.}

\item{plot}{Logical. If \code{TRUE} (default), a plot is printed.}
}
\value{
A numeric matrix of GATES generic target estimates
}
\description{
Accessor function for the GATES generic target estimates after VEIN
}
