% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunization.R
\docType{data}
\name{immunization}
\alias{immunization}
\title{Immunization data}
\format{
An object of class \code{data.frame} with 843 rows and 38 columns.
}
\usage{
immunization
}
\description{
TODO: add description
}
\keyword{datasets}
