% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{initialize_diff}
\alias{initialize_diff}
\title{Initializer for the arguments \code{diff_GATES} and \code{diff_CLAN} for the \code{GenericML()} function.}
\usage{
initialize_diff(subtract_from = "most", subtracted = 1)
}
\arguments{
\item{subtract_from}{String indicating the base group to subtract from, either \code{"most"} (default) or \code{"least"}. The most affected group corresponds to the \eqn{K}-th group in the paper (there are \eqn{K} groups). The least affected group corresponds to the first group.}

\item{subtracted}{Vector indicating the groups to be subtracted from the group specified in \code{subtract_from}. If there are \eqn{K} group, \code{subtracted} should be a subset of \eqn{{1,2,...,K}}.}
}
\description{
Returns a list with two elements called \code{subtract_from} and \code{subtracted}. The first element (\code{subtract_from}) denotes what shall be the base group to subtract from in the generic targets of interest (GATES or CLAN); either \code{"most"} or \code{"least"}. The second element (\code{subtracted}) are the groups to be subtracted from \code{subtract_from}, which is a subset of \eqn{{1,2,...,K}}, where \eqn{K} equals the number of groups. The number of groups should be consistent with the number of groups induced by the argument \code{quantile_cutoffs}.
}
