% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lambda_parameters}
\alias{lambda_parameters}
\title{Returns the two lambda parameters that are used to find the best ML method.}
\usage{
lambda_parameters(BLP, GATES, proxy_CATE, membership)
}
\arguments{
\item{BLP}{An instance of the class \code{BLP}.}

\item{GATES}{An instance of the class \code{GATES}.}

\item{proxy_CATE}{Proxy estimates of the CATE.}

\item{membership}{A logical matrix that indicates the group membership of each observation. Needs to be an instance of \code{\link{quantile_group}}.}
}
\value{
A list containing the parameters \code{lambda} and \code{lambda.bar}.
}
\description{
Returns the two lambda parameters that are used to find the best ML method.
}
