% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedeffects.R
\name{make_fixedeffects}
\alias{make_fixedeffects}
\title{Make fixed effects identifiers}
\usage{
make_fixedeffects(...)
}
\arguments{
\item{...}{Vectors of group memberships}
}
\value{
A vector of unique fixed-effect identifiers
}
\description{
Create a vector of possibly multiway fixed-effects identifiers for each observation. Expects group membership ID variables as input. Passing \code{k} vectors of group memberships will create fixed-effect idenifiers for \code{k}-way fixed effects.
The returned vector may be passed as input to \code{\link{GenericML}()}.
}
\examples{

## create 2-way fixed effects (cluster-time) for 5 observations
cluster <- c(1, 1, 2, 2, 1)
time <- c("time1", "time1", "time2", "time1", "time2")
make_fixedeffects(cluster, time)

}
