% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcredit.R
\docType{data}
\name{microcredit}
\alias{microcredit}
\title{Morocco microcredit data}
\format{
An object of class \code{data.frame} with 5513 rows and 23 columns.
}
\usage{
microcredit
}
\description{
TODO: add description
}
\keyword{datasets}
