% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.GenericML}
\alias{plot.GenericML}
\title{Makes plots for generic ML}
\usage{
\method{plot}{GenericML}(
  x,
  learner = "best",
  type = "GATES",
  CLAN_variable = NULL,
  groups = "all",
  limits = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{An instance of \code{\link{GenericML}}.}

\item{learner}{The learner whose results are to be returned. Default is \code{"best"} for the best learner as measured by the lambda parameters.}

\item{type}{The analysis to be plotted. Either \code{"GATES"}, \code{"BLP"}, or \code{"CLAN"}. Default is \code{"GATES"}.}

\item{CLAN_variable}{Name of CLAN variable to be plotted. Only applicable if \code{type = "CLAN"}.}

\item{groups}{Character vector for the groups to be plotted for GATES and CLAN. Default is \code{"all"} for all groups. If there are \eqn{K} groups, this variable is a subset of \code{c("G1", "G2",...,"GK", "GK-G1", "GK-G2",...)}, but this set depends on the choices of the arguments \code{"diff_GATES"} and \code{"diff_CLAN"} of the \code{\link{GenericML}} function that generated the instance \code{x}.}

\item{limits}{The limits of the y-axis of the plot.}

\item{title}{The title of the plot.}

\item{...}{Additional arguments.}
}
\description{
Makes plots for generic ML
}
