% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.BLP_info}
\alias{print.BLP_info}
\title{Print method for a \code{"BLP_info"} object}
\usage{
\method{print}{BLP_info}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of the class \code{"BLP_info"}, as returned by the function \code{\link{get_BLP}()}.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments to be passed down.}
}
\value{
A print to the console.
}
\description{
Print method for a \code{"BLP_info"} object
}
