% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.GATES_info}
\alias{print.GATES_info}
\title{Print method for a \code{"GATES_info"} object}
\usage{
\method{print}{GATES_info}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of the class \code{"GATES_info"}, as returned by the function \code{\link{get_GATES}()}.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments to be passed down.}
}
\value{
A print to the console.
}
\description{
Print method for a \code{"GATES_info"} object
}
