% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.heterogeneity_CLAN}
\alias{print.heterogeneity_CLAN}
\title{Print method for a \code{"heterogeneity_CLAN"} object}
\usage{
\method{print}{heterogeneity_CLAN}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"\link{heterogeneity_CLAN}"}.}

\item{...}{Additional arguments to be passed down.}
}
\value{
A print to the console.
}
\description{
Print method for a \code{"heterogeneity_CLAN"} object
}
