% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{propensity.score}
\alias{propensity.score}
\title{Estimates the propensity score}
\usage{
propensity.score(Z, D, estimator = "constant")
}
\arguments{
\item{Z}{a matrix or data frame of covariates}

\item{D}{a binary vector of treatment status}

\item{estimator}{the estimator to be used. Either a numeric vector (which is then taken as estimates of the propensity scores) or a string specifying the estimator. The string must either be equal to 'constant' (estimates the propensity scores by mean(D)), 'elastic.net', 'random.forest', 'tree', or mlr3 syntax. Example for the latter: mlr3::lrn("classif.ranger", num.trees = 500) for a classification forest.}
}
\value{
Estimates of \eqn{Pr(D = 1 | Z)} and an 'mlr3' object of the employed model (if applicable)
}
\description{
Estimates the propensity score
}
