% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{propensity_score}
\alias{propensity_score}
\title{Estimates the propensity score}
\usage{
propensity_score(Z, D, estimator = "constant")
}
\arguments{
\item{Z}{A matrix of covariates}

\item{D}{A binary vector of treatment status}

\item{estimator}{The estimator for the propensity scores. Either a numeric vector (which is then taken as estimates of the propensity scores) or a string specifying the estimator. The string must either be equal to \code{'constant'} (estimates the propensity scores by \code{mean(D)}), \code{'elastic_net'}, \code{'random_forest'}, \code{'tree'}, or \code{mlr3} syntax. Note that in case of \code{mlr3} syntax, do \emph{not} specify if the learner is a regression learner or classification learner; Example: \code{'mlr3::lrn("ranger", num.trees = 500)'} for a random forest learner. Note that this is a string and the absence of the \code{classif.} or \code{regr.} keywords.}
}
\value{
Estimates of \eqn{Pr(D = 1 | Z)} and an \code{mlr3} object of the employed model (if applicable).
}
\description{
Estimates the propensity score
}
