% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{setup_vcov}
\alias{setup_vcov}
\title{Setup function for the arguments \code{vcov_BLP} and \code{vcov_GATES} in the \code{\link{GenericML}} function.}
\usage{
setup_vcov(estimator = "vcovHC", arguments = list(type = "const"))
}
\arguments{
\item{estimator}{String specifying a covariance matrix estimator in the \href{https://cran.r-project.org/web/packages/sandwich/}{sandwich}. Default is \code{"vcovHC"}. Recommended estimators are \code{"vcovBS"}, \code{"vcovCL"}, \code{"vcovHAC"}, and \code{"vcovHC"}.}

\item{arguments}{A list of arguments that are to be passed to the function in the sandwich package that is specified in \code{estimator}. Default is \code{list(type = "const")}, which specifies the homoskedastic ordinary least squares covariance matrix estimate.}
}
\description{
Returns a list with two elements called \code{estimator} and \code{arguments}. The element \code{estimator} is a string specifying the covariance matrix estimator to be used in the linear regression regression of interest and needs to be a covariance estimator function in the \href{https://cran.r-project.org/web/packages/sandwich/}{sandwich} package. The second element, \code{arguments}, is a list of arguments that shall be passed to the function specified in the first element, \code{estimator}.
}
