% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializers.R
\name{setup_vcov_defaultFE}
\alias{setup_vcov_defaultFE}
\title{Setup covariance estimator with fixed effects}
\usage{
setup_vcov_defaultFE(fixed_effects = NULL)
}
\arguments{
\item{fixed_effects}{A vector of fixed-effect group identifiers; possibly \code{NULL} for no fixed effects (default)}
}
\description{
This is a simple wrapper for function \code{\link{setup_vcov}()} when fixed effects are involved. Please refer to that function's documentation for details on the behavior. If \code{fixed_effects = NULL}, then the standard homoskedastic least squares covariance estimator will be used. If \code{fixed_effects} is not \code{NULL} but a vector of fixed-effect group membership identifiers, then a cluster-robust covariance matrix estimator will be used where we cluster on the fixed effects identifiers. Depending on the nature of your fixed effects (especially when working with panel data), it may be preferable to cluster on other variables. Such alternative specifications can be made with the wrapped function \code{\link{setup_vcov}()}.
}
