% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stratified}
\alias{stratified}
\title{performs stratified sampling}
\usage{
stratified(group, relative.size, select = NULL)
}
\arguments{
\item{group}{The column or columns that should be used to create the groups. Can be a character vector of column names (recommended) or a numeric vector of column.}

\item{relative.size}{proportional to the number of observations per group.}

\item{select}{A named list containing levels from the "group" variables in which you are interested. The list names must be present as variable names for the input dataset.

#################################################
The function below effectively does the following:
DF <- data.frame(}
}
\description{
performs stratified sampling
}
