% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{TrueIfUnix}
\alias{TrueIfUnix}
\title{A function that returns \code{TRUE} if your OS is a Unix system and \code{FALSE} otherwise.}
\usage{
TrueIfUnix()
}
\description{
A function that returns \code{TRUE} if your OS is a Unix system and \code{FALSE} otherwise.
}
