% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blp.R
\name{BLP}
\alias{BLP}
\title{Estimates the BLP parameters based on the main sample M.}
\usage{
BLP(
  D,
  Y,
  propensity_scores,
  proxy.baseline,
  proxy.cate,
  HT = FALSE,
  X1_control = setup_X1(),
  vcov_control = setup_vcov(),
  significance_level = 0.05
)
}
\arguments{
\item{D}{a binary vector of treatment status of length \emph{|M|}}

\item{Y}{a vector of responses of length \emph{|M|}}

\item{propensity_scores}{a vector of propensity scores of length \emph{|M|}}

\item{proxy.baseline}{a vector of proxy baseline estimates of length \emph{M}}

\item{proxy.cate}{a vector of proxy CATE estimates of length \emph{M}}

\item{HT}{logical. If TRUE, a HT transformation is applied (BLP2 in the paper). Default is FALSE.}

\item{X1_control}{Specifies the design matrix \eqn{X_1} in the regression. See the documentation of \code{\link{setup_X1}} for details.}

\item{vcov_control}{Specifies the covariance matrix estimator. See the documentation of \code{\link{setup_vcov}} for details.}

\item{significance_level}{significance level for construction of confidence intervals}
}
\value{
BLP coefficients with inference statements
}
\description{
Estimates the BLP parameters based on the main sample M.
}
