% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blp.R
\name{BLP}
\alias{BLP}
\title{Performs BLP regression.}
\usage{
BLP(
  Y,
  D,
  propensity_scores,
  proxy_baseline,
  proxy_CATE,
  HT = FALSE,
  X1_control = setup_X1(),
  vcov_control = setup_vcov(),
  significance_level = 0.05
)
}
\arguments{
\item{Y}{a vector of responses.}

\item{D}{a binary vector of treatment status.}

\item{propensity_scores}{a vector of propensity scores.}

\item{proxy_baseline}{a vector of proxy baseline estimates.}

\item{proxy_CATE}{a vector of proxy CATE estimates.}

\item{HT}{logical. If \code{TRUE}, a HT transformation is applied (BLP2 in the paper). Default is \code{FALSE}.}

\item{X1_control}{Specifies the design matrix \eqn{X_1} in the regression. See the documentation of \code{\link{setup_X1}} for details.}

\item{vcov_control}{Specifies the covariance matrix estimator. See the documentation of \code{\link{setup_vcov}} for details.}

\item{significance_level}{significance level for construction of confidence intervals}
}
\value{
An object of the class "BLP".
}
\description{
Performs the linear regression for the "best linear predictor" procedure.
}
