% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blp.R
\name{BLP}
\alias{BLP}
\title{Estimates the BLP parameters based on the main sample M.}
\usage{
BLP(
  D,
  Y,
  propensity.scores,
  proxy.baseline,
  proxy.cate,
  HT.transformation = FALSE,
  X1.variables = list(functions_of_Z = c("B"), custom_covariates = NULL, fixed_effects
    = NULL),
  vcov.control = list(estimator = "vcovHC", arguments = list(type = "const")),
  significance.level = 0.05
)
}
\arguments{
\item{D}{a binary vector of treatment status of length \emph{|M|}}

\item{Y}{a vector of responses of length \emph{|M|}}

\item{propensity.scores}{a vector of propensity scores of length \emph{|M|}}

\item{proxy.baseline}{a vector of proxy baseline estimates of length \emph{M}}

\item{proxy.cate}{a vector of proxy CATE estimates of length \emph{M}}

\item{HT.transformation}{logical. If TRUE, a HT transformation is applied (BLP2 in the paper). Default is FALSE.}

\item{X1.variables}{a list controlling the variables that shall be used in the matrix X1. The first element of the list, functions_of_Z, needs to be a subset of c("S", "B", "p"), where "p" corresponds to the propensity scores (default is "B"). The seconds element, custom_covariates, is an optional matrix/data frame of custom covariates that shall be included in X1 (default is NULL). The third element, fixed_effects, is a vector of integers that indicates group membership of the observations: For each group, a fixed effect will be added (default is NULL for no fixed effects). Note that in the final matrix X1, a constant 1 will be silently included so that the regression model has an intercept.}

\item{vcov.control}{a list with two elements called 'estimator' and 'arguments'. The argument 'estimator' is a string specifying the covariance matrix estimator to be used; specifies a covariance estimator function in the sandwich package (https://cran.r-project.org/web/packages/sandwich/sandwich.pdf). Recommended estimators are "vcovBS", "vcovCL", "vcovHAC", and "vcovHC". Default is 'vcovHC'. The element 'arguments' is a list of arguments that shall be passed to the function specified in the element 'estimator'. Default leads to the (homoskedastic) ordinary least squares covariance matrix estimator. See the reference manual of the sandwich package for details (https://cran.r-project.org/web/packages/sandwich/vignettes/sandwich.pdf).}

\item{significance.level}{significance level for construction of confidence intervals}
}
\value{
BLP coefficients with inference statements
}
\description{
Estimates the BLP parameters based on the main sample M.
}
