% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clan.R
\name{CLAN}
\alias{CLAN}
\title{Estimates the CLAN parameters in the main sample}
\usage{
CLAN(
  Z_CLAN.main.sample,
  group.membership.main.sample,
  equal.group.variances = FALSE,
  differences.control = list(group.to.subtract.from = "most", groups.to.be.subtracted =
    1),
  significance.level = 0.05
)
}
\arguments{
\item{Z_CLAN.main.sample}{a matrix with \emph{|M|} rows. Each column represents a variable for which CLAN shall be performed.}

\item{group.membership.main.sample}{a logical matrix with \emph{M} rows that indicate
the group memberships (such a matrix is returned by the function quantile.group())}

\item{equal.group.variances}{logical. If TRUE, the the two within-group variances of the most and least affected group are assumed to be equal. Default is FALSE.}

\item{differences.control}{a list with two elements called 'group.to.subtract.from' and 'groups.to.be.subtracted'. The first element ('group.to.subtract.from') denotes what shall be the base group to subtract from in CLAN; either "most" or "least". The second element ('groups.to.be.subtracted') are the groups to be subtracted from 'group.to.subtract.from', which is a subset of {1,...,K}, where K equals the number of groups.}

\item{significance.level}{Significance level. Default is 0.05.}
}
\value{
The two CLAN parameters ("most" affected and "least" affected) for each variable in Z_CLAN.main.sample
}
\description{
Estimates the CLAN parameters in the main sample
}
