% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clan.R
\name{CLAN}
\alias{CLAN}
\title{Estimates the CLAN parameters in the main sample}
\usage{
CLAN(
  Z_CLAN.main.sample,
  group.membership.main.sample,
  equal.group.variances = FALSE,
  diff = setup_diff(),
  significance_level = 0.05
)
}
\arguments{
\item{Z_CLAN.main.sample}{a matrix with \emph{|M|} rows. Each column represents a variable for which CLAN shall be performed.}

\item{group.membership.main.sample}{a logical matrix with \emph{M} rows that indicate
the group memberships (such a matrix is returned by the function quantile_group())}

\item{equal.group.variances}{logical. If TRUE, the the two within-group variances of the most and least affected group are assumed to be equal. Default is FALSE.}

\item{diff}{Controls the generic targets of CLAN. See the documentation of \code{\link{setup_diff}}.}

\item{significance_level}{Significance level. Default is 0.05.}
}
\value{
The two CLAN parameters ("most" affected and "least" affected) for each variable in Z_CLAN.main.sample
}
\description{
Estimates the CLAN parameters in the main sample
}
