% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{GATES}
\alias{GATES}
\title{Estimates the GATES parameters based on the main sample M.}
\usage{
GATES(
  D,
  Y,
  propensity_scores,
  proxy.baseline,
  proxy.cate,
  group.membership.main.sample,
  HT = FALSE,
  X1_control = setup_X1(),
  vcov_control = setup_vcov(),
  diff = setup_diff(),
  significance_level = 0.05
)
}
\arguments{
\item{D}{a binary vector of treatment status of length \emph{|M|}}

\item{Y}{a vector of responses of length \emph{|M|}}

\item{propensity_scores}{a vector of propensity scores of length \emph{|M|}}

\item{proxy.baseline}{a vector of proxy baseline estimates of length \emph{M}}

\item{proxy.cate}{a vector of proxy CATE estimates of length \emph{M}}

\item{group.membership.main.sample}{a logical matrix with \emph{M} rows that indicate
the group memberships (such a matrix is returned by the function quantile_group())}

\item{HT}{logical. If TRUE, a HT transformation is applied (GATES2 in the paper). Default is FALSE.}

\item{X1_control}{Specifies the design matrix \eqn{X_1} in the regression. See the documentation of \code{\link{setup_X1}} for details.}

\item{vcov_control}{Specifies the covariance matrix estimator. See the documentation of \code{\link{setup_vcov}} for details.}

\item{diff}{Controls the generic targets of GATES See the documentation of \code{\link{setup_diff}}.}

\item{significance_level}{significance level for construction of confidence intervals}
}
\description{
Estimates the GATES parameters based on the main sample M.
}
