% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{GATES}
\alias{GATES}
\title{Estimates the GATES parameters based on the main sample M.}
\usage{
GATES(
  D,
  Y,
  propensity.scores,
  proxy.baseline,
  proxy.cate,
  group.membership.main.sample,
  HT = FALSE,
  X1.variables = list(functions_of_Z = c("B"), custom_covariates = NULL, fixed_effects
    = NULL),
  vcov_control = setup_vcov(),
  diff = setup_diff(),
  significance_level = 0.05
)
}
\arguments{
\item{D}{a binary vector of treatment status of length \emph{|M|}}

\item{Y}{a vector of responses of length \emph{|M|}}

\item{propensity.scores}{a vector of propensity scores of length \emph{|M|}}

\item{proxy.baseline}{a vector of proxy baseline estimates of length \emph{M}}

\item{proxy.cate}{a vector of proxy CATE estimates of length \emph{M}}

\item{group.membership.main.sample}{a logical matrix with \emph{M} rows that indicate
the group memberships (such a matrix is returned by the function quantile.group())}

\item{HT}{logical. If TRUE, a HT transformation is applied (GATES2 in the paper). Default is FALSE.}

\item{X1.variables}{a list controlling the variables that shall be used in the matrix X1. The first element of the list, functions_of_Z, needs to be a subset of c("S", "B", "p"), where "p" corresponds to the propensity scores (default is "B"). The seconds element, custom_covariates, is an optional matrix/data frame of custom covariates that shall be included in X1 (default is NULL). The third element, fixed_effects, is a vector of integers that indicates group membership of the observations: For each group, a fixed effect will be added (default is NULL for no fixed effects). Note that in the final matrix X1, a constant 1 will be silently included so that the regression model has an intercept.}

\item{vcov_control}{Specifies the covariance matrix estimator. See the documentation of \code{\link{setup_vcov}} for details.}

\item{diff}{Controls the generic targets of GATES See the documentation of \code{\link{setup_diff}}.}

\item{significance_level}{significance level for construction of confidence intervals}
}
\description{
Estimates the GATES parameters based on the main sample M.
}
