% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get.generic.ml.for.given.learner}
\alias{get.generic.ml.for.given.learner}
\title{Performs generic ML for a given learning technique (with only one split of the data)}
\usage{
get.generic.ml.for.given.learner(
  Z,
  D,
  Y,
  propensity.scores,
  learner = "mlr3::lrn(\\"cv_glmnet\\", s = \\"lambda.min\\")",
  M.set,
  A.set,
  Z_CLAN = NULL,
  X1_BLP = list(functions_of_Z = c("B"), custom_covariates = NULL, fixed_effects =
    NULL),
  X1_GATES = list(functions_of_Z = c("B"), custom_covariates = NULL, fixed_effects =
    NULL),
  HT = FALSE,
  vcov_BLP = list(estimator = "vcovHC", arguments = list(type = "const")),
  vcov_GATES = list(estimator = "vcovHC", arguments = list(type = "const")),
  equal_variances_CLAN = FALSE,
  quantile_cutoffs = c(0.25, 0.5, 0.75),
  diff_GATES = list(group.to.subtract.from = "most", groups.to.be.subtracted = 1),
  diff_CLAN = list(group.to.subtract.from = "most", groups.to.be.subtracted = 1),
  significance_level = 0.05,
  min_variation = 1e-05
)
}
\arguments{
\item{Z}{a matrix or data frame of covariates}

\item{D}{a binary vector of treatment status of length}

\item{Y}{a vector of responses of length}

\item{propensity.scores}{a vector of propensity scores}

\item{learner}{The machine learner that shall be used}

\item{M.set}{main set}

\item{A.set}{auxiliary set}

\item{Z_CLAN}{A matrix of variables that shall be considered for the CLAN. If \code{NULL} (default), then \code{Z_CLAN = Z}, i.e. CLAN is performed for all variables in \code{Z}.}

\item{X1_BLP}{a list controlling the variables that shall be used in the matrix X1 for the BLP regression. The first element of the list, functions_of_Z, needs to be a subset of c("S", "B", "p"), where "p" corresponds to the propensity scores (default is "B"). The second element, custom_covariates, is an optional matrix/data frame of custom covariates that shall be included in X1 (default is NULL). The third element, fixed_effects, is a vector of integers that indicates group membership of the observations: For each group, a fixed effect will be added (default is NULL for no fixed effects). Note that in the final matrix X1, a constant 1 will be silently included so that the regression model has an intercept.}

\item{X1_GATES}{a list controlling the variables that shall be used in the matrix X1 for the GATES regression. The first element of the list, functions_of_Z, needs to be a subset of c("S", "B", "p"), where "p" corresponds to the propensity scores (default is "B"). The second element, custom_covariates, is an optional matrix/data frame of custom covariates that shall be included in X1 (default is NULL). The third element, fixed_effects, is a vector of integers that indicates group membership of the observations: For each group, a fixed effect will be added (default is NULL for no fixed effects). Note that in the final matrix X1, a constant 1 will be silently included if no HT transformation is applied so that the regression model has an intercept.}

\item{HT}{logical. If TRUE, a HT transformation is applied in BLP and GATES. Default is FALSE.}

\item{vcov_BLP}{a list with two elements called 'estimator' and 'arguments'. The argument 'estimator' is a string specifying the covariance matrix estimator to be used in the BLP regression; specifies a covariance estimator function in the sandwich package (https://cran.r-project.org/web/packages/sandwich/sandwich.pdf). Recommended estimators are "vcovBS", "vcovCL", "vcovHAC", and "vcovHC". Default is 'vcovHC'. The element 'arguments' is a list of arguments that shall be passed to the function specified in the element 'estimator'. Default leads to the (homoskedastic) ordinary least squares covariance matrix estimator. See the reference manual of the sandwich package for details (https://cran.r-project.org/web/packages/sandwich/vignettes/sandwich.pdf).}

\item{vcov_GATES}{same as vcov_BLP, just for GATES regression}

\item{equal_variances_CLAN}{logical. If TRUE, the the two within-group variances of the most and least affected group in CLAN are assumed to be equal. Default is FALSE.}

\item{quantile_cutoffs}{Cutoff points of quantiles that shall be used for GATES grouping}

\item{diff_GATES}{a list with two elements called 'group.to.subtract.from' and 'groups.to.be.subtracted'. The first element ('group.to.subtract.from') denotes what shall be the base group to subtract from in GATES; either "most" or "least". The second element ('groups.to.be.subtracted') are the groups to be subtracted from 'group.to.subtract.from', which is a subset of {1,...,K}, where K equals the number of groups.}

\item{diff_CLAN}{same as diff_GATES, just for CLAN.}

\item{significance_level}{Significance level. Default is 0.05}

\item{min_variation}{minimum variation of the predictions before random noise with distribution N(0, var(Y)/20) is added. Default is 1e-05.

TODO: instructions on how mlr3 input is supposed to work (needs to be a string!)
TODO: comments on CLAN: If there are categorical variables, apply one-hot-encoding to Z_CLAN. The interpretation then becomes: Is there a factor that is overproportionally present in the least or most affected group?}
}
\description{
Performs generic ML for a given learning technique (with only one split of the data)
}
