% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lambda_parameters}
\alias{lambda_parameters}
\title{returns the two parameters that are used to find the best ML method}
\usage{
lambda_parameters(BLP.obj, GATES.obj, proxy_CATE.M_set, membership)
}
\arguments{
\item{BLP.obj}{an object as returned by CLAN()}

\item{GATES.obj}{an object as returned by get.BLP.parameters()}

\item{proxy_CATE.M_set}{Proxy CATE estimators for the main sample}

\item{membership}{a logical matrix with \emph{M} rows that indicate
the group memberships (such a matrix is returned by the function quantile_group())}
}
\value{
lambda and lambda.bar parameters
}
\description{
returns the two parameters that are used to find the best ML method
}
