% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Med}
\alias{Med}
\title{Find lower and upper median of a vector.}
\usage{
Med(x)
}
\arguments{
\item{x}{A numeric vector whose medians we are interested in.}
}
\value{
A list with the upper and lower median and the Med statistic.
}
\description{
See Comment 4.2 in the paper.
}
