% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Med}
\alias{Med}
\title{Calculate lower and upper median}
\usage{
Med(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with the upper and lower median and the Med statistic (which is their mean).
}
\description{
Calculates the lower and and median of a vector as proposed in Comment 4.2 in the paper.
}
\examples{
set.seed(1)
x <- runif(100)
Med(x)

}
\references{
Chernozhukov V., Demirer M., Duflo E., Fernández-Val I. (2020). \dQuote{Generic Machine Learning Inference on Heterogenous Treatment Effects in Randomized Experiments.} \emph{arXiv preprint arXiv:1712.04802}. URL: \url{https://arxiv.org/abs/1712.04802}.

Lang M., Binder M., Richter J., Schratz P., Pfisterer F., Coors S., Au Q., Casalicchio G., Kotthoff L., Bischl B. (2019). \dQuote{mlr3: A Modern Object-Oriented Machine Learning Framework in R.} \emph{Journal of Open Source Software}, \bold{4}(44), 1903. \doi{10.21105/joss.01903}.
}
