% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.GenericML}
\alias{plot.GenericML}
\title{Makes plots for generic ML}
\usage{
\method{plot}{GenericML}(
  x,
  learner = "best",
  type = "GATES",
  CLAN.variable = NULL,
  groups.to.be.plotted = "all",
  limits = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{an object as returned by GenericML()}

\item{learner}{the learner whose results are to be returned. Default is 'best'}

\item{type}{the analysis to be plotted. Either 'GATES', 'BLP', or 'CLAN'. Default is 'GATES'.}

\item{CLAN.variable}{CLAN variable to be plotted. Only applicable if type = 'CLAN'.}

\item{groups.to.be.plotted}{The groups to be plotted for GATES and CLAN. Default is 'all'. If there are K groups, this variable can be set to a subset of {"G1", "G2",...,"GK", "GK-G1", "GK-G2",...}, but this set depends on the choices of the arguments 'differences.control_GATES' and 'differences.control_CLAN' of the GenericML() function that generated GenericML.obj.}

\item{limits}{the limits of the y-axis of the plot.}

\item{title}{the title of the plot.}

\item{...}{Additional arguments}
}
\description{
Makes plots for generic ML
}
