% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.GenericML}
\alias{print.GenericML}
\title{prints information on a GenericML instance}
\usage{
\method{print}{GenericML}(x, ...)
}
\arguments{
\item{x}{an object returned by \emph{GenericML()}}

\item{...}{additional arguments to be passed}
}
\description{
prints information on a GenericML instance
}
