% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.GenericML}
\alias{print.GenericML}
\title{prints information on an instance of \code{\link{GenericML}}.}
\usage{
\method{print}{GenericML}(x, ...)
}
\arguments{
\item{x}{An instance of \code{\link{GenericML}}.}

\item{...}{additional arguments to be passed}
}
\description{
prints information on an instance of \code{\link{GenericML}}.
}
