% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{proxy_BCA}
\alias{proxy_BCA}
\title{Baseline Conditional Average}
\usage{
proxy_BCA(Z, D, Y, A_set, learner, min_variation = 1e-05)
}
\arguments{
\item{Z}{A numeric design matrix that holds the covariates in its columns.}

\item{D}{A binary vector of treatment assignment. Value one denotes assignment to the treatment group and value zero assignment to the control group.}

\item{Y}{A numeric vector containing the response variable.}

\item{A_set}{a numerical vector of indices of observations in the auxiliary sample. The}

\item{learner}{A string specifying the machine learner for the estimation. Either \code{'elastic_net'}, \code{'random_forest'}, \code{'tree'}, or a custom learner specified with \code{mlr3} syntax. In the latter case, do \emph{not} specify in the \code{mlr3} syntax specification if the learner is a regression learner or classification learner. Example: \code{'mlr3::lrn("ranger", num.trees = 100)'} for a random forest learner with 100 trees. Note that this is a string and the absence of the \code{classif.} or \code{regr.} keywords. See \url{https://mlr3learners.mlr-org.com} for a list of \code{mlr3} learners.}

\item{min_variation}{Specifies a threshold for the minimum variation of the predictions. If the variation of a BCA prediction falls below this threshold, random noise with distribution \eqn{N(0, var(Y)/20)} is added to it. Default is \code{1e-05}.}
}
\value{
An object of class \code{proxy_BCA}, consisting of the following components:
\describe{
\item{\code{estimates}}{A numeric vector of BCA estimates of each observation.}
\item{\code{mlr3_objects}}{\code{mlr3} objects used for estimation.}
}
}
\description{
Proxy estimation of the Baseline Conditional Average (BCA), defined by \eqn{E[Y | D=0, Z]}. Estimation is done on the auxiliary sample, while BCA predictions are made for all observations.
}
\references{
Chernozhukov, V., Demirer, M., Duflo, E., and Fernández-Val, I. (2021). Generic Machine Learning Inference on Heterogenous Treatment Effects in Randomized Experiments. \href{https://arxiv.org/abs/1712.04802}{\emph{arXiv preprint arXiv:1712.04802}}.
}
\seealso{
\code{\link{proxy_CATE}}
}
