% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{proxy_BCA}
\alias{proxy_BCA}
\title{Performs estimation of the baseline conditional average (BCA), \eqn{E[Y | D=0, Z]}, on the auxiliary sample}
\usage{
proxy_BCA(Z, D, Y, A_set, learner = "random.forest", min_variation = 1e-05)
}
\arguments{
\item{Z}{an ( \emph{n} x \emph{d}) matrix or data frame of covariates}

\item{D}{a binary vector of treatment status of length \emph{n}}

\item{Y}{a vector of responses of length \emph{n}}

\item{A_set}{a numerical vector of indices of observations in the auxiliary sample.}

\item{learner}{the classification machine learner to be used. Either 'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_g = mlr3::lrn("regr.ranger", num.trees = 500) for a regression forest, which is also the default.}

\item{min_variation}{minimum variation of the predictions before random noise with distribution N(0, var(Y)/20) is added. Default is 1e-05.}
}
\value{
An object of the class \code{proxy_BCA}.
}
\description{
Performs estimation of the baseline conditional average (BCA), \eqn{E[Y | D=0, Z]}, on the auxiliary sample
}
