% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{proxy_CATE}
\alias{proxy_CATE}
\title{Performs estimation of the conditional average treatment effect (CATE), \eqn{E[Y | D=1, Z] - E[Y | D=0, Z]}, on the auxiliary sample}
\usage{
proxy_CATE(
  Z,
  D,
  Y,
  A_set,
  learner = "random.forest",
  proxy_BCA = NULL,
  min_variation = 1e-05
)
}
\arguments{
\item{Z}{an ( \emph{n} x \emph{d}) matrix or data frame of covariates}

\item{D}{a binary vector of treatment status of length \emph{n}}

\item{Y}{a vector of responses of length \emph{n}}

\item{A_set}{a numerical vector of indices of observations in the auxiliary sample.}

\item{learner}{the regression machine learner to be used. Either 'glm', 'random.forest', or 'tree'. Can alternatively be specified by using the mlr3 framework, for example ml_g = mlr3::lrn("regr.ranger", num.trees = 500) for a regression forest, which is also the default.}

\item{proxy_BCA}{A vector of length \emph{n} of proxy estimates of the baseline estimator \eqn{E[Y | D=0, Z]}. If NULL, these will be estimated separately.}

\item{min_variation}{minimum variation of the predictions before random noise with distribution N(0, var(Y)/20) is added. Default is 1e-05.}
}
\value{
An object of the class \code{proxy_CATE}.
}
\description{
Performs estimation of the conditional average treatment effect (CATE), \eqn{E[Y | D=1, Z] - E[Y | D=0, Z]}, on the auxiliary sample
}
