% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3.R
\name{proxy_CATE}
\alias{proxy_CATE}
\title{Performs estimation of the conditional average treatment effect (CATE), \eqn{E[Y | D=1, Z] - E[Y | D=0, Z]}, on the auxiliary sample}
\usage{
proxy_CATE(
  Z,
  D,
  Y,
  A_set,
  learner = "random_forest",
  proxy_BCA = NULL,
  min_variation = 1e-05
)
}
\arguments{
\item{Z}{A matrix or data frame of the covariates.}

\item{D}{A binary vector of treatment assignment.}

\item{Y}{The response vector.}

\item{A_set}{a numerical vector of indices of observations in the auxiliary sample.}

\item{learner}{A string specifying the machine learner to be used for estimation. Either \code{'elastic_net'}, \code{'random_forest'} (default), or \code{'tree'}. Can alternatively be specified by using \code{mlr3} syntax \emph{without} specification if the learner is a regression learner or classification learner. Example: \code{'mlr3::lrn("ranger", num.trees = 500)'} for a random forest learner. Note that this is a string and the absence of the \code{classif.} or \code{regr.} keywords. See \url{https://mlr3learners.mlr-org.com} for a list of \code{mlr3} learners.}

\item{proxy_BCA}{A vector of proxy estimates of the baseline estimator BCA, \eqn{E[Y | D=0, Z]}. If \code{NULL}, these will be estimated separately.}

\item{min_variation}{Minimum variation of the predictions before random noise with distribution \eqn{N(0, var(Y)/20)} is added. Default is \code{1e-05}.}
}
\value{
An object of the class \code{proxy_CATE}.
}
\description{
Performs estimation of the conditional average treatment effect (CATE), \eqn{E[Y | D=1, Z] - E[Y | D=0, Z]}, on the auxiliary sample
}
