% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{quantile_group}
\alias{quantile_group}
\title{Partition a vector x into groups based on its quantiles}
\usage{
quantile_group(x, cutoffs = c(0.25, 0.5, 0.75), quantile.nam = TRUE)
}
\arguments{
\item{x}{the vector to be partitioned}

\item{cutoffs}{the quantile cutoffs for the partition. Default are the quartiles.}

\item{quantile.nam}{logical. Shall the cutoff values be included?}
}
\description{
Partition a vector x into groups based on its quantiles
}
