% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-checks.R
\name{TrueIfUnix}
\alias{TrueIfUnix}
\title{Check if user's OS is a Unix system}
\usage{
TrueIfUnix()
}
\value{
A Boolean that is \code{TRUE} if the user's operating system is a Unix system and \code{FALSE} otherwise.
}
\description{
Check if user's OS is a Unix system
}
